refactor: 상수 추출
