--- src/domain/constants/card.ts
@@ -1,10 +1,0 @@
-export const MAIN_COLOR = {
-  BC카드: '#E83F44',
-  신한카드: '#0046FF',
-  카카오카드: '#FFE300',
-  현대카드: '#000000',
-  우리카드: '#0283CA',
-  롯데카드: '#E60013',
-  하나카드: '#009692',
-  국민카드: '#776C61',
-};
